# .stainless/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Graph Italia API",
    "description": "API documentation for the Graph Italia application",
    "version": "1.0.0"
  },
  "components": {
    "securitySchemes": { "bearerAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT" } }
  },
  "security": [{ "bearerAuth": [] }],
  "servers": [
    { "url": "http://localhost:3000", "description": "Local server" },
    { "url": "http://graph-test.italia.it", "description": "Staging server" },
    { "url": "http://graph.italia.it", "description": "Production server" }
  ],
  "paths": {
...
    "/api/charts": {
      "get": {
        "operationId": "getApiCharts",
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": { "type": "string" },
                      "userId": { "type": "string" },
                      "name": { "anyOf": [{ "type": "string" }, { "type": "null" }] },
                      "description": { "anyOf": [{ "type": "string" }, { "type": "null" }] },
                      "chart": { "type": "string" },
                      "config": { "anyOf": [{}, { "type": "null" }] },
                      "data": { "anyOf": [{}, { "type": "null" }] },
                      "isRemote": { "anyOf": [{ "type": "boolean" }, { "type": "null" }] },
                      "remoteUrl": { "anyOf": [{ "type": "string" }, { "type": "null" }] },
                      "publish": { "type": "boolean" },
                      "createdAt": { "type": "string" },
                      "updatedAt": { "type": "string" }
                    },
                    "required": [
                      "id",
                      "userId",
                      "name",
                      "description",
                      "chart",
                      "config",
                      "data",
                      "isRemote",
                      "remoteUrl",
                      "publish",
                      "createdAt",
                      "updatedAt"
                    ]
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": { "error": { "type": "string" } },
                  "required": ["error"]
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": { "error": { "type": "string" } },
                  "required": ["error"]
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "postApiCharts",
        "responses": {
          "201": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": { "type": "string" },
                    "userId": { "type": "string" },
                    "name": { "anyOf": [{ "type": "string" }, { "type": "null" }] },
                    "description": { "anyOf": [{ "type": "string" }, { "type": "null" }] },
                    "chart": { "type": "string" },
                    "config": { "anyOf": [{}, { "type": "null" }] },
                    "data": { "anyOf": [{}, { "type": "null" }] },
                    "isRemote": { "anyOf": [{ "type": "boolean" }, { "type": "null" }] },
                    "remoteUrl": { "anyOf": [{ "type": "string" }, { "type": "null" }] },
                    "publish": { "type": "boolean" },
                    "createdAt": { "type": "string" },
                    "updatedAt": { "type": "string" }
                  },
                  "required": [
                    "id",
                    "userId",
                    "name",
                    "description",
                    "chart",
                    "config",
                    "data",
                    "isRemote",
                    "remoteUrl",
                    "publish",
                    "createdAt",
                    "updatedAt"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": { "error": { "type": "string" } },
                  "required": ["error"]
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": { "error": { "type": "string" } },
                  "required": ["error"]
                }
              }
            }
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string" },
                  "description": { "type": "string" },
                  "chart": { "type": "string" },
                  "config": {},
                  "data": {},
                  "remoteUrl": { "anyOf": [{ "type": "string" }, { "type": "null" }] },
                  "isRemote": { "type": "boolean" },
                  "publish": { "type": "boolean" },
                  "preview": { "anyOf": [{ "type": "string" }, { "type": "null" }] }
                },
                "required": ["chart"]
              }
            }
          }
        }
      }
    },
...
    "/api/hints": {
      "post": {
        "operationId": "postApiHints",
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "hint": { "type": "string" },
                      "confidence": { "type": "number", "minimum": 0, "maximum": 1 }
                    },
                    "required": ["hint", "confidence"]
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": { "error": { "type": "string" } },
                  "required": ["error"]
                }
              }
            }
          }
        }
      }
    }
  }
}